Add files via upload
This is the basic structure of the files that will be representing our web site, created and designed by Trill.  We will have all of this stuff ready for the assignment due date on the 22nd August.  Sections as of yet might be copy-pasted from work in Microsoft Teams or added incrementally during the course of the project.
Each section will be changed to reflect who is meant to be doing which part.